+++ src/articles/dto/create-article.dto.ts
@@ -1,16 +1,7 @@
-import { ApiProperty, OmitType, PartialType, PickType } from "@nestjs/swagger";
+import { ApiProperty } from "@nestjs/swagger";
 import { IsNotEmpty } from "class-validator";
 
-export class ArticleDto {
-  @ApiProperty({
-    example: "1",
-    description: "文章的 ID",
-  })
-  @IsNotEmpty({
-    message: "id 為必填欄位。",
-  })
-  public readonly id: string;
-
+export class CreateArticleDto {
   @ApiProperty({
     description: "文章標題",
     example: "我是第一篇文章",
@@ -38,9 +29,3 @@ export class ArticleDto {
   })
   public readonly contents: string;
 }
-
-export class CreateArticleDto extends OmitType(ArticleDto, ["id"] as const) {}
-
-export class DeleteArticleDto extends PickType(ArticleDto, ["id"] as const) {}
-
-export class UpdateArticleDto extends PartialType(CreateArticleDto) {}
