+++ src/articles/dto/create-comment.dto.ts
@@ -1,7 +1,7 @@
-import { ApiProperty, PickType } from "@nestjs/swagger";
+import { ApiProperty } from "@nestjs/swagger";
 import { IsNotEmpty } from "class-validator";
 
-export class CommentDto {
+export class CreateCommentDto {
   @ApiProperty({
     description: "某文章的留言訊息",
     example: "我是第一篇文章的留言",
@@ -11,7 +11,3 @@ export class CommentDto {
   })
   public readonly contents: string;
 }
-
-export class CreateCommentDto extends PickType(CommentDto, [
-  "contents",
-] as const) {}
